Support for opening doors in the map
Converted all the tiles into objects, so we can keep state of stuff like if the door is open etc.

When a player interacts we try to open in door in the grid tiles directly surrounding the player.

Animate doors should animate the opening and closing of the doors in the near future.